Trait massa_execution_exports::controller_traits::ExecutionController

pub trait ExecutionController: Send + Sync {
    // Required methods
    fn update_blockclique_status(
        &self,
        finalized_blocks: HashMap<Slot, BlockId>,
        new_blockclique: Option<HashMap<Slot, BlockId>>,
        block_storage: PreHashMap<BlockId, Storage>
    );
    fn get_filtered_sc_output_event(
        &self,
        filter: EventFilter
    ) -> Vec<SCOutputEvent> ;
    fn get_final_and_candidate_balance(
        &self,
        addresses: &[Address]
    ) -> Vec<(Option<Amount>, Option<Amount>)> ;
    fn get_op_exec_status(
        &self
    ) -> (HashMap<OperationId, bool>, HashMap<OperationId, bool>);
    fn get_final_and_active_data_entry(
        &self,
        input: Vec<(Address, Vec<u8>)>
    ) -> Vec<(Option<Vec<u8>>, Option<Vec<u8>>)> ;
    fn get_cycle_active_rolls(&self, cycle: u64) -> BTreeMap<Address, u64>;
    fn execute_readonly_request(
        &self,
        req: ReadOnlyExecutionRequest
    ) -> Result<ReadOnlyExecutionOutput, ExecutionError>;
    fn unexecuted_ops_among(
        &self,
        ops: &PreHashSet<OperationId>,
        thread: u8
    ) -> PreHashSet<OperationId>;
    fn get_addresses_infos(
        &self,
        addresses: &[Address]
    ) -> Vec<ExecutionAddressInfo> ;
    fn get_stats(&self) -> ExecutionStats;
    fn clone_box(&self) -> Box<dyn ExecutionController>;
}

interface that communicates with the execution worker thread

Required Methods

fn update_blockclique_status( &self, finalized_blocks: HashMap<Slot, BlockId>, new_blockclique: Option<HashMap<Slot, BlockId>>, block_storage: PreHashMap<BlockId, Storage> )

Updates blockclique status by signaling newly finalized blocks and the latest blockclique.

Arguments

finalized_blocks: newly finalized blocks indexed by slot.

blockclique: new blockclique (if changed). Indexed by slot.

block_storage: storage instances for new blocks. Each one owns refs to the block and its ops/endorsements/parents.

fn get_filtered_sc_output_event(&self, filter: EventFilter) -> Vec<SCOutputEvent>

Get execution events optionally filtered by:

start slot

end slot

emitter address

original caller address

operation id

fn get_final_and_candidate_balance( &self, addresses: &[Address] ) -> Vec<(Option<Amount>, Option<Amount>)>

Get the final and active values of balance.

Return value

(final_balance, active_balance)

fn get_op_exec_status( &self ) -> (HashMap<OperationId, bool>, HashMap<OperationId, bool>)

Get the execution status of operation that have been executed both speculatively or finaly

Return value (speculative_statuses, final_statuses) for each hashmap: key: the operation id value: true: operation executed successfully, false: operation failed

fn get_final_and_active_data_entry( &self, input: Vec<(Address, Vec<u8>)> ) -> Vec<(Option<Vec<u8>>, Option<Vec<u8>>)>

Get a copy of a single datastore entry with its final and active values

Return value

(final_data_entry, active_data_entry)

fn get_cycle_active_rolls(&self, cycle: u64) -> BTreeMap<Address, u64>

Returns for a given cycle the stakers taken into account by the selector. That correspond to the roll_counts in cycle - 3.

By default it returns an empty map.

fn execute_readonly_request( &self, req: ReadOnlyExecutionRequest ) -> Result<ReadOnlyExecutionOutput, ExecutionError>

Execute read-only SC function call without causing modifications to the consensus state

arguments

req: an instance of ReadOnlyCallRequest describing the parameters of the execution

returns

An instance of ExecutionOutput containing a summary of the effects of the execution, or an error if the execution failed.

fn unexecuted_ops_among( &self, ops: &PreHashSet<OperationId>, thread: u8 ) -> PreHashSet<OperationId>

List which operations inside the provided list were not executed

fn get_addresses_infos(&self, addresses: &[Address]) -> Vec<ExecutionAddressInfo>

Gets information about a batch of addresses

fn get_stats(&self) -> ExecutionStats

Get execution statistics

fn clone_box(&self) -> Box<dyn ExecutionController>

Returns a boxed clone of self. Useful to allow cloning Box<dyn ExecutionController>.