Trait massa_execution_exports::ExecutionController

pub trait ExecutionController: Send + Sync {
    fn update_blockclique_status(
        &self,
        finalized_blocks: HashMap<Slot, BlockId>,
        blockclique: HashMap<Slot, BlockId>
    );
    fn get_filtered_sc_output_event(
        &self,
        filter: EventFilter
    ) -> Vec<SCOutputEvent>;
    fn get_final_and_active_parallel_balance(
        &self,
        addresses: Vec<Address>
    ) -> Vec<(Option<Amount>, Option<Amount>)>;
    fn get_final_and_active_data_entry(
        &self,
        input: Vec<(Address, Vec<u8>)>
    ) -> Vec<(Option<Vec<u8>>, Option<Vec<u8>>)>;
    fn get_every_final_datastore_key(&self, addr: &Address) -> Vec<Vec<u8>>;
    fn execute_readonly_request(
        &self,
        req: ReadOnlyExecutionRequest
    ) -> Result<ExecutionOutput, ExecutionError>;
    fn clone_box(&self) -> Box<dyn ExecutionController>;
}

interface that communicates with the execution worker thread

Required Methods

fn update_blockclique_status(
&self,
finalized_blocks: HashMap<Slot, BlockId>,
blockclique: HashMap<Slot, BlockId>
)

Updates blockclique status by signaling newly finalized blocks and the latest blockclique.

Arguments

finalized_blocks: newly finalized blocks

blockclique: new blockclique

fn get_filtered_sc_output_event(&self, filter: EventFilter) -> Vec<SCOutputEvent>

Get execution events optionally filtered by:

start slot

end slot

emitter address

original caller address

operation id

fn get_final_and_active_parallel_balance(
&self,
addresses: Vec<Address>
) -> Vec<(Option<Amount>, Option<Amount>)>

Get a balance final and active values

Return value

(final_balance, active_balance)

fn get_final_and_active_data_entry(
&self,
input: Vec<(Address, Vec<u8>)>
) -> Vec<(Option<Vec<u8>>, Option<Vec<u8>>)>

Get a copy of a single datastore entry with its final and active values

Return value

(final_data_entry, active_data_entry)

fn get_every_final_datastore_key(&self, addr: &Address) -> Vec<Vec<u8>>

Get every datastore key of the given address.

Returns

A vector containing all the keys

fn execute_readonly_request(
&self,
req: ReadOnlyExecutionRequest
) -> Result<ExecutionOutput, ExecutionError>

Execute read-only SC function call without causing modifications to the consensus state

arguments

req: an instance of ReadOnlyCallRequest describing the parameters of the execution

returns

An instance of ExecutionOutput containing a summary of the effects of the execution, or an error if the execution failed.

fn clone_box(&self) -> Box<dyn ExecutionController>

Returns a boxed clone of self. Useful to allow cloning Box<dyn ExecutionController>.