Trait massa_execution_exports::ExecutionManager

pub trait ExecutionManager {
    fn stop(&mut self);
}

Execution manager used to stop the execution thread

Required methods

fn stop(&mut self)

Stop the execution thread Note that we do not take self by value to consume it because it is not allowed to move out of Box This will improve if the unsized_fn_params feature stabilizes enough to be safely usable.