massa_execution_exports

Module types_trace_info

types for execution-trace

General description

‘execution-trace’ is a Rust feature that can be enabled for the execution module to provide trace of coin transfers.

Note: that the ‘execution-trace’ feature is also present in massa-sc-runtime.

Transfers collected

TX transfers: Massa coins transferred by regular operation.

SC transfers: Massa coins transferred by various ABI calls.

Transfers are nested (if happened during a SC call or localCall).

Note: in massa-sc-runtime, all SC ABI are collected and stored (see assembly_script_get_balance Note 2: there is a filter in get_slot_transfers functions (json-rpc get_slots_transfers & grpc get_slot_transfers) to return only the coin transfers.

API

json-rpc public api:

get_slots_transfers

grpc public api:

get_slot_transfers

grpc public stream api:

new_slot_transfers

new_slot_abi_call_stacks

Usage

The ‘execution-trace’ feature was originally developed for the transfer-indexer which is used as a starting point for exchange (CEX) integration.

Tests

Unit tests in massa-execution-worker

Type Aliases

ExecutionResult

ABI and execution succeed or not