Trait massa_execution_exports::ExecutionController

pub trait ExecutionController: Send + Sync {
    fn update_blockclique_status(
        &self, 
        finalized_blocks: Map<BlockId, Block>, 
        blockclique: Map<BlockId, Block>
    );

    fn get_filtered_sc_output_event(
        &self, 
        start: Option<Slot>, 
        end: Option<Slot>, 
        emitter_address: Option<Address>, 
        original_caller_address: Option<Address>, 
        original_operation_id: Option<OperationId>
    ) -> Vec<SCOutputEvent>;

    fn get_final_and_active_ledger_entry(
        &self, 
        addr: &Address
    ) -> (Option<LedgerEntry>, Option<LedgerEntry>);

    fn execute_readonly_request(
        &self, 
        req: ReadOnlyExecutionRequest
    ) -> Result<ExecutionOutput, ExecutionError>;

    fn clone_box(&self) -> Box<dyn ExecutionController>;
}

interface that communicates with the execution worker thread

Required methods

fn update_blockclique_status(
&self,
finalized_blocks: Map<BlockId, Block>,
blockclique: Map<BlockId, Block>
)

Updates blockclique status by signalling newly finalized blocks and the latest blockclique.

arguments

finalized_blocks: newly finalized blocks

blockclique: new blockclique

fn get_filtered_sc_output_event(
&self,
start: Option<Slot>,
end: Option<Slot>,
emitter_address: Option<Address>,
original_caller_address: Option<Address>,
original_operation_id: Option<OperationId>
) -> Vec<SCOutputEvent>

Get execution events optionnally filtered by:

start slot

end slot

emitter address

original caller address

operation id

fn get_final_and_active_ledger_entry(
&self,
addr: &Address
) -> (Option<LedgerEntry>, Option<LedgerEntry>)

Get a copy of a full ledger entry with its final and active values

return value

(final_entry, active_entry)

fn execute_readonly_request(
&self,
req: ReadOnlyExecutionRequest
) -> Result<ExecutionOutput, ExecutionError>

Execute read-only bytecode without causing modifications to the consensus state

arguments

req: an instance of ReadOnlyExecutionRequest describing the parameters of the execution

returns

An instance of ExecutionOutput containing a summary of the effects of the execution, or an error if the execution failed.

fn clone_box(&self) -> Box<dyn ExecutionController>

Returns a boxed clone of self. Useful to alow cloning Box.