Module massa_execution_exports::types

This file exports useful types used to interact with the execution worker

Structs

ExecutedBlockInfo

structure storing a block id + network versions (from a block header)

ExecutionAddressInfo

Execution info about an address

ExecutionBlockMetadata

Metadata needed to execute the block

ExecutionOutput

structure describing the output of a single execution

ExecutionQueryCycleInfos

Information about cycles

ExecutionQueryRequest

Request to atomically execute a batch of execution state queries

ExecutionQueryResponse

Response to a list of execution queries

ExecutionQueryStakerInfo

Staker information for a given cycle

ExecutionStackElement

Structure describing an element of the execution stack. Every time a function is called from bytecode, a new ExecutionStackElement is pushed at the top of the execution stack to represent the local execution context of the called function, instead of the caller’s which should lie just below in the stack.

ReadOnlyCallRequest

structure describing a read-only call

ReadOnlyExecutionOutput

structure describing the output of a read only execution

ReadOnlyExecutionRequest

structure describing different types of read-only execution request

Enums

ExecutionQueryExecutionStatus

Execution status of an operation or denunciation

ExecutionQueryRequestItem

Execution state query item

ExecutionQueryResponseItem

Execution state query response item

ReadOnlyExecutionTarget

structure describing different possible targets of a read-only execution request

SlotExecutionOutput

structure describing the output of the execution of a slot